// src/components/profile-details/types.ts
export type FormFields = {
  first_name: string;
  last_name: string;
  bio: string;
  years_of_experience: number;
  location: string;
  skills: string;
  tools_used: string;
  keywords: string;
  fields_of_interest: string;
  linkedin_url: string;
  github_url: string;
  website_url: string;
  position: string;
  company_id: string;
  school_id: string;
  academic_major_id: string;
  highest_degree: string;
};